chore: split portfinder from socket-server
