+++ src/hmr/socket-server.js
@@ -1,24 +1,9 @@
 const http = require('http')
 const WebSocket = require('ws')
 const express = require('express')
-const portfinder = require('portfinder')
-const { log, error } = require('../utils/logger')
+const { clog } = require('../utils/logger')
 
-// eslint-disable-next-line consistent-return
-const getFreePort = async (port) => {
-  try {
-    return await portfinder.getPortPromise({
-      port, // minimum port
-      stopPort: port + 1000, // maximum port
-    })
-  } catch (e) {
-    error('[ERROR]: Something went wrong when trying to find a free port for HMR server!')
-    error(e)
-  }
-}
-
-const startServer = async (_port) => {
-  const port = await getFreePort(_port)
+const startServer = (port) => {
   const app = express()
 
   app.get('/', (req, res) => {
@@ -37,10 +22,10 @@ const startServer = async (_port) => {
   // })
 
   server.listen(port, () => {
-    log(`HMR Server running on port: ${server.address().port}`)
+    clog('\n[@pitcher/watcher]:', 'white', `HMR Server running on port: ${server.address().port}`, 'cyan')
   })
 
-  return { port, server, wss }
+  return { server, wss }
 }
 
 module.exports = startServer
